Factor out suggestions logic into its own trait
